fix(grading): full PSA half-grade support end-to-end
PSA does issue half grades (9.5, 8.5, 7.5, 6.5, 5.5, 4.5, 3.5, 2.5);
the canonical label maps only had 1.5. Expand the PSA maps in all four
places that mirror each other so half-grade slabs render with proper
short-form labels (e.g. 'NM-MT+ 8.5', 'EX-MT+ 6.5') instead of a bare
'PSA 8.5' fallback:

- server/src/utils/grade-labels.ts  (source of truth, AI scanner uses it)
- client/src/lib/grade-labels.ts    (client mirror)
- client/src/pages/SubReturns.tsx   gradeLabel() and COMPANY_LABEL_MAP

Plus one correctness fix in catalog.service.ts reassignCatalogRow:
the SQL filter casted grade to ::int, which would silently drop any
half-grade slab when reassigning a catalog row (8.5 -> 8, miss match,
slab stranded under old catalog). Cast to ::numeric(4,1) to match the
column type.

No DB migration; slab_details.grade is already numeric(4,1).

+++ server/src/services/catalog.service.ts
@@ -645,7 +645,7 @@ export async function reassignCatalogRow(userId: string, params: {
       AND ci.catalog_id = ${params.old_catalog_id}
       AND LOWER(TRIM(COALESCE(ci.card_name_override, ''))) = LOWER(TRIM(${params.card_name ?? ''}))
       AND sd.company = ${params.company}
-      AND sd.grade IS NOT DISTINCT FROM ${params.grade}::int
+      AND sd.grade IS NOT DISTINCT FROM ${params.grade}::numeric(4,1)
   `.execute(db);
   return { updated: Number(result.numAffectedRows ?? 0) };
 }
